requirements.txt
jsonschema
PyJWT
cffi==1.11.5
cryptography==2.2.2
pyOpenSSL==18.0.0
